requirements.txt
fastapi
uvicorn[standard]
qdrant-client
openai
pymupdf
python-dotenv
whisper
open_clip_torch
torch
Pillow
